Moved code to proper folders
