+++ src/app/Caches/Query/QueryEntries.tsx
@@ -18,10 +18,10 @@ import {
   ToolbarItemVariant
 } from '@patternfly/react-core';
 import {SearchIcon} from '@patternfly/react-icons';
-import cacheService from '../../services/cacheService';
+import cacheService from '../../../services/cacheService';
 import {Table, TableBody, TableHeader, TableVariant} from '@patternfly/react-table';
 import {TableErrorState} from "@app/Common/TableErrorState";
-import displayUtils from "../../services/displayUtils";
+import displayUtils from "../../../services/displayUtils";
 import SyntaxHighlighter from 'react-syntax-highlighter';
 import {githubGist} from 'react-syntax-highlighter/dist/esm/styles/hljs';
